Refactored the connection page such that it also works in the Felix WebConsole
