+++ flexiblepower.ui.connectionspage/bnd.bnd
@@ -1,5 +1,7 @@
 -buildpath: ${default-buildpath},\
 	org.apache.felix.http.servlet-api,\
+	org.apache.felix.webconsole,\
+	com.google.gson,\
 	flexiblepower.ral;version=latest,\
 	flexiblepower.runtime;version=latest,\
 	flexiblepower.ui;version=latest
@@ -7,5 +9,7 @@
 Bundle-Version: 1.0.0.${qualifier}
 Private-Package: org.flexiblepower.runtime.ui.connectionspage
 Service-Component: org.flexiblepower.runtime.ui.connectionspage.*
-Import-Package:  \
+Import-Package: \
+	org.apache.felix.webconsole;resolution:=optional,\
+	org.flexiblepower.ui;resolution:=optional,\
 	*
