+++ flexiblepower.ui.connectionspage/res/widgets/DashboardWidget/dom.jsPlumb-1.7.5.js
