+++ flexiblepower.ui.connectionspage/res/widgets/DashboardWidget/menu.png
