+++ flexiblepower.ui.connectionspage/res/widgets/DashboardWidget/style.css
@@ -1,15 +1,14 @@
-<style type="text/css">
-#main {
+#graph {
 	user-select: none;
 	-moz-user-select: none;
 	-khtml-user-select: none;
 	-webkit-user-select: none;
-	-ms-user-select: none;
+	-ms-user-select: none;	
 	position: absolute;
-	left: 0;
-	width: 80%;
 	top: 150px;
 	bottom: 130px;
+	left: 0px;
+	right: 320px;
 	padding: 1em;
 	overflow: auto;
 	background: white;
@@ -21,7 +20,7 @@
 	bottom: 130px;
 	right: 0;
 	padding: 1em;
-	width: 20%;
+	width: 300px;
 	overflow: auto;
 	background: white;
 	border-left: 2px solid #33333;
@@ -75,31 +74,40 @@
 	font-size: 70%;
 }
 
-#main ._jsPlumb_connector {
+#graph ._jsPlumb_connector {
     z-index: 4;
 }
 
-#main ._jsPlumb_endpoint {
+#graph ._jsPlumb_endpoint {
     z-index: 22;
     cursor: pointer;
     opacity: 0.5;
 }
 
-#main ._jsPlumb_endpoint._jsPlumb_hover {
+#graph ._jsPlumb_endpoint._jsPlumb_hover {
 	opacity: 0.7;
 }
 
-#main ._jsPlumb_overlay {
+#graph ._jsPlumb_overlay {
     z-index: 21;
     font-weight: bold !important;
 }
-</style>
 
-<div id="main">
-</div>
-<div id="properties">
-	<input type="button" value="Autoconnect" id="autoconnect" />
-	<h1>Properties of <span id="componentname">&lt;no selection&gt;</span></h1>
-</div>
-<script src="dom.jsPlumb-1.7.5.js"></script>
-<script src="script.js"></script>
+/* Specific stuff for the felix webconsole layout */
+.ui-widget #graph {
+	top: 120px;
+	left: 96px;
+	right: 416px;
+	bottom: 0;
+}
+
+.ui-widget #properties {
+	top: 120px;
+	right: 96px;
+	width: 300px;
+	bottom: 0;
+}
+
+#navmenu li ul {
+	z-index: 30;
+}
